Cleanup

+++ Bachelor-Thesis/src/Core/Chunk.cpp
@@ -2,14 +2,14 @@
 
 namespace Lang {
 	
-	void Chunk::Write(uint16_t byte, uint16_t line) {
+	void Chunk::Write(uint8_t byte, uint16_t line) {
 		Code.push_back(byte);
 		Lines.push_back(line);
 	}
 
-	uint16_t Chunk::AddConstant(Value value) {
+	uint8_t Chunk::AddConstant(Value value) {
 		Constants.push_back(value);
-		return (uint16_t)(Constants.size() - 1);
+		return (uint8_t)(Constants.size() - 1);
 	}
 
 }
