+++ Bachelor-Thesis/src/Core/Chunk.h
@@ -17,16 +17,16 @@ namespace Lang {
 	};
 
 	struct Chunk {
-		std::vector<uint16_t> Code;
+		std::vector<uint8_t> Code;
 		std::vector<uint16_t> Lines;
 		std::vector<Value> Constants;
 
 		/// <summary>Writes a byte to the bytcode.</summary>
-		void Write(uint16_t byte, uint16_t line);
+		void Write(uint8_t byte, uint16_t line);
 
 		/// <summary>Adds a new value to the constants array.</summary>
 		/// <returns>The index at which the value was added.</returns>
-		uint16_t AddConstant(Value value);
+		uint8_t AddConstant(Value value);
 	};
 
 }
